src/requirements.txt
gunicorn
Werkzeug==1.0.1
matplotlib==3.4.2
numpy==1.19.5
flask==1.1.2
Pillow==8.4.0
scikit_learn==1.0.1
tensorflow==2.7.0
